refactor: split codec out of type modules, move tests to tests/
Separate what types are from how they go on the wire, and align the
crate with the gateway test layout (all tests in `tests/`, zero inline
tests in `src/`, matching `jmux-proto` and `devolutions-gateway`).

- `src/control_codec.rs` (new): `TAG_*` constants, `impl Encode /
  Decode for ControlMessage`, plus the private `encode_subnets /
  decode_subnets / encode_domains / decode_domains /
  encode_cert_renewal_result / decode_cert_renewal_result` helpers.
- `src/control.rs`: now just type definitions, constructors, and the
  `protocol_version()` accessor. Reads as a pure data contract.
- `src/session_codec.rs` (new): `TAG_*` constants + `impl Encode /
  Decode for ConnectRequest / ConnectResponse`.
- `src/session.rs`: types + constructors + `is_success` /
  `protocol_version` only.
- `src/lib.rs`: registers the two new `pub(crate)` codec modules and
  re-exports `CertRenewalResult` at the crate root (needed by the new
  external test file).
- `tests/control.rs` (new): the six `DomainName::matches_hostname`
  tests, nine `ControlMessage` roundtrip tests, the oversized-message
  rejection test, and the proptest — all moved out of `src/control.rs`
  so the production file no longer has a trailing `#[cfg(test)]`
  block. Uses only the public API.
- `tests/session_roundtrip.rs` → `tests/session.rs` (git rename) for
  naming symmetry with `tests/control.rs`.

Tests still green: 3 (version) + 16 (control integration) +
5 (session integration) on agent-tunnel-proto; 20 on agent-tunnel and
84 on devolutions-gateway lib untouched. Zero clippy warnings; fmt
clean.
